fix(llm): cast token counts to i64 for correct OTel trace attribute types
tracing-opentelemetry 0.33's SpanAttributeVisitor only implements
record_i64 (not record_u64). When u32/u64 values are passed to
span.record(), they fall through to record_debug and get exported
as String("100") instead of I64(100) in OTLP span attributes.

This caused OTel/Langfuse trace data to have token usage fields
(input_tokens, output_tokens, cache_read/creation_input_tokens)
and max_tokens exported as strings rather than integers, making
them unusable for numeric aggregation in trace backends.

Fix: cast all u32 token counts and max_tokens to i64 before calling
span.record() in both Anthropic and OpenAI providers.

Added tests in trace_field_export.rs that:
- Verify i64 values export as I64 (correct)
- Document the u32→String bug (regression guard)
- Verify full deferred-record pattern with context_activation

+++ crates/visp-llm/src/anthropic.rs
@@ -501,7 +501,7 @@ impl LlmProvider for AnthropicProvider {
             langfuse.trace.metadata = field::Empty,
         );
         span.record("gen_ai.system", "anthropic");
-        span.record("gen_ai.request.max_tokens", config.max_tokens);
+        span.record("gen_ai.request.max_tokens", config.max_tokens as i64);
         span.record("gen_ai.request.temperature", config.temperature);
 
         // Langfuse trace-level fields: record when enabled
@@ -737,22 +737,25 @@ fn byte_stream_to_chat_events(
                 };
 
                 // 在 span 上 record usage / model / finish_reasons / cost
+                // Cast u32 → i64: tracing-opentelemetry's Visit impl only handles
+                // record_i64 (not record_u64), so u32/u64 values would fall through
+                // to record_debug and be exported as String("100") instead of I64(100).
                 state
                     .span
-                    .record("gen_ai.usage.input_tokens", state.input_tokens);
+                    .record("gen_ai.usage.input_tokens", state.input_tokens as i64);
                 state
                     .span
-                    .record("gen_ai.usage.output_tokens", state.output_tokens);
+                    .record("gen_ai.usage.output_tokens", state.output_tokens as i64);
                 if state.cache_read_input_tokens > 0 {
                     state.span.record(
                         "gen_ai.usage.cache_read.input_tokens",
-                        state.cache_read_input_tokens,
+                        state.cache_read_input_tokens as i64,
                     );
                 }
                 if state.cache_creation_input_tokens > 0 {
                     state.span.record(
                         "gen_ai.usage.cache_creation.input_tokens",
-                        state.cache_creation_input_tokens,
+                        state.cache_creation_input_tokens as i64,
                     );
                 }
                 if state.stop_reason == "max_tokens" {
@@ -1638,7 +1641,7 @@ mod tests {
             visp.llm.cost_usd = tracing::field::Empty,
             visp.llm.token_limit_hit = tracing::field::Empty,
         );
-        span.record("gen_ai.request.max_tokens", 4096u64);
+        span.record("gen_ai.request.max_tokens", 4096i64);
         span.record("gen_ai.request.temperature", 0.7f64);
 
         drop(_guard);
@@ -1673,7 +1676,7 @@ mod tests {
 
         // 模拟 chat_stream 中在 span 创建后的 record
         span.record("gen_ai.system", "anthropic");
-        span.record("gen_ai.request.max_tokens", 4096u64);
+        span.record("gen_ai.request.max_tokens", 4096i64);
         span.record("gen_ai.request.temperature", 0.7f64);
 
         drop(_guard);
@@ -1792,7 +1795,7 @@ mod tests {
             visp.llm.cost_usd = tracing::field::Empty,
             visp.llm.token_limit_hit = tracing::field::Empty,
         );
-        span.record("gen_ai.request.max_tokens", 4096u64);
+        span.record("gen_ai.request.max_tokens", 4096i64);
         span.record("gen_ai.request.temperature", 0.7f64);
 
         let sse = make_complete_sse("claude-sonnet-4-6", "end_turn");
